UPDATED SMS WITH ALERTS

+++ SMS_PROJECT_NEW/UI/SMS-PROJECT/src/app/components/course/course.component.ts
@@ -1,6 +1,7 @@
 import { Component, OnInit } from '@angular/core';
 import { Router, ActivatedRoute } from '@angular/router';
 import { CourseserviceService } from 'src/app/courseservice.service';
+import Swal from 'sweetalert2'
 
 @Component({
   selector: 'app-course',
@@ -102,16 +103,45 @@ export class CourseComponent implements OnInit {
   
 
   deleteCourse(data) {
-    this.courseserviceService.deleteCourse(data)
-      .subscribe(
-        data => {
-          console.log(data);
-          
-        },
-        error => console.log(error));
-        alert("Deleted")
-        this.reloadData();
+    Swal.fire({
+      title: 'Are you sure?',
+      text: "You won't be able to revert this!",
+      icon: 'warning',
+      showCancelButton: true,
+      confirmButtonColor: '#3085d6',
+      cancelButtonColor: '#d33',
+      confirmButtonText: 'Yes, delete it!'
+    }).then((result) => {
+      if (result.value) {
+        this.courseserviceService.deleteCourse(data)
+        .subscribe(
+          data => {
+            console.log(data);
+            
+          },
+          error => console.log(error));
+          Swal.fire({
+            position: 'center',
+            icon: 'success',
+            title: 'Deleted',
+            showConfirmButton: true,
+            timer: 5500
+          })
+          this.reloadData();
+      }
+    })
   }
+  // deleteCourseSure(data) {
+  //   this.courseserviceService.deleteCourse(data)
+  //     .subscribe(
+  //       data => {
+  //         console.log(data);
+          
+  //       },
+  //       error => console.log(error));
+  //       alert("Deleted")
+  //       this.reloadData();
+  // }
  
  }
