+++ SMS_PROJECT_NEW/UI/SMS-PROJECT/src/app/components/students/students.component.ts
@@ -74,7 +74,17 @@ export class StudentsComponent implements OnInit {
   
 
   deleteStudent(data) {
-    this.studentsService.deleteStudent(data)
+    Swal.fire({
+      title: 'Are you sure?',
+      text: "You won't be able to revert this!",
+      icon: 'warning',
+      showCancelButton: true,
+      confirmButtonColor: '#3085d6',
+      cancelButtonColor: '#d33',
+      confirmButtonText: 'Yes, delete it!'
+    }).then((result) => {
+      if (result.value) {
+        this.studentsService.deleteStudent(data)
       .subscribe(
         data => {
           console.log(data);
@@ -90,6 +100,9 @@ export class StudentsComponent implements OnInit {
           timer: 5500
         })
         this.reloadData();
+      }
+    })
+    
   }
  
  }
